Version 1.3.1
Repair od full text TOC upload to BIB records - if the record was to long (more than 45000 bytes), it was not trimmed, but sent to upload to database by manage-18. This caused to removal of some other BIB fields at the end of the records. Although this was reported by email, this needed manual repair of the record. This bug has been repaired.